+++ src/main/scala/ai/tripl/arc/extract/AvroExtract.scala
@@ -263,7 +263,6 @@ object AvroExtractStage {
       }
     }
 
-    if (!arcContext.isStreaming) repartitionedDF.rdd.setName(stage.outputView)
     if (arcContext.immutableViews) repartitionedDF.createTempView(stage.outputView) else repartitionedDF.createOrReplaceTempView(stage.outputView)
 
     if (!repartitionedDF.isStreaming) {
@@ -272,7 +271,7 @@ object AvroExtractStage {
       stage.stageDetail.put("numPartitions", java.lang.Integer.valueOf(repartitionedDF.rdd.partitions.length))
 
       if (stage.persist) {
-        repartitionedDF.persist(arcContext.storageLevel)
+        spark.catalog.cacheTable(stage.outputView, arcContext.storageLevel)
         stage.stageDetail.put("records", java.lang.Long.valueOf(repartitionedDF.count))
       }
     }