+++ src/main/scala/ai/tripl/arc/extract/BytesExtract.scala
@@ -179,7 +179,6 @@ object BytesExtractStage {
       case None => df
     }
 
-    if (!arcContext.isStreaming) repartitionedDF.rdd.setName(stage.outputView)
     if (arcContext.immutableViews) repartitionedDF.createTempView(stage.outputView) else repartitionedDF.createOrReplaceTempView(stage.outputView)
 
     stage.stageDetail.put("inputFiles", java.lang.Integer.valueOf(repartitionedDF.inputFiles.length))
@@ -187,7 +186,7 @@ object BytesExtractStage {
     stage.stageDetail.put("numPartitions", java.lang.Integer.valueOf(repartitionedDF.rdd.partitions.length))
 
     if (stage.persist && !repartitionedDF.isStreaming) {
-      repartitionedDF.persist(arcContext.storageLevel)
+      spark.catalog.cacheTable(stage.outputView, arcContext.storageLevel)
       stage.stageDetail.put("records", java.lang.Long.valueOf(repartitionedDF.count))
     }