+++ src/main/scala/ai/tripl/arc/extract/HTTPExtract.scala
@@ -296,14 +296,13 @@ object HTTPExtractStage {
       }
     }
 
-    if (!arcContext.isStreaming) repartitionedDF.rdd.setName(stage.outputView)
     if (arcContext.immutableViews) repartitionedDF.createTempView(stage.outputView) else repartitionedDF.createOrReplaceTempView(stage.outputView)
 
     stage.stageDetail.put("outputColumns", java.lang.Integer.valueOf(repartitionedDF.schema.length))
     stage.stageDetail.put("numPartitions", java.lang.Integer.valueOf(repartitionedDF.rdd.partitions.length))
 
     if (stage.persist) {
-      repartitionedDF.persist(arcContext.storageLevel)
+      spark.catalog.cacheTable(stage.outputView, arcContext.storageLevel)
 
       val records = repartitionedDF.count
       stage.stageDetail.put("records", java.lang.Long.valueOf(records))