+++ src/main/scala/ai/tripl/arc/extract/ImageExtract.scala
@@ -167,7 +167,6 @@ object ImageExtractStage {
       }
     }
 
-    if (!arcContext.isStreaming) repartitionedDF.rdd.setName(stage.outputView)
     if (arcContext.immutableViews) repartitionedDF.createTempView(stage.outputView) else repartitionedDF.createOrReplaceTempView(stage.outputView)
 
     if (!repartitionedDF.isStreaming) {
@@ -176,7 +175,7 @@ object ImageExtractStage {
       stage.stageDetail.put("numPartitions", Integer.valueOf(repartitionedDF.rdd.partitions.length))
 
       if (stage.persist) {
-        repartitionedDF.persist(arcContext.storageLevel)
+        spark.catalog.cacheTable(stage.outputView, arcContext.storageLevel)
         stage.stageDetail.put("records", java.lang.Long.valueOf(repartitionedDF.count))
       }
     }